Refactor: move middleware definition to a separate module
